bindata: CPU request for migrator should not be higher than average use
The migrator is always running and does not consume 100mc, which is against
our policy for out of the box components. The pod is still burstable
which does not prevent excess CPU from being consumed, but components are
not allowed to significantly overstate their steady state CPU. A future
alternative is for the operator to only run the migrator when necessary,
in which case the CPU and memory use can be increased to be consistent
with use during runtime.