README: Stub in brief purpose docs
Would be nice to have the README include (links to) collaboration
notes and security issue reporting and all that.  But I'm not clear on
what the operator maintainers want to use for that stuff, so this
commit just brings over and tweaks a bit of text from [1] so folks
wondering "what is this about?" don't have to think too hard.

[1]: https://github.com/openshift/kubernetes-kube-storage-version-migrator/blob/481bd04dbc78c29dad05deba0309e049fa26e4d9/USER_GUIDE.md#storage-version-migrator-in-a-nutshell